update 5:57pm
Added an if statement that makes ti to where if all fields are empty it will show a text that says "Please fill out all fields".
Wrote an else statement with an if statement that if the email is less than 3 characters it will send an alert.
Wrote an else statement that if it's longer than 3 characters it will lead to another else statement to see if the email address is proper and if not then it will say"Please use valid email.
If the email is valid then it will check if the password matches with the password in the confirm password box.
If everything is correct it well say "name+' is registereed with email ' + email;"